requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
duckduckgo-search>=4.1.0
fastapi>=0.110.0
uvicorn>=0.27.0  # For running the FastAPI server
